+++ app/assets/javascripts/app/views/aspect_membership_blueprint_view.js
@@ -1,5 +1,3 @@
-//= require ./aspects_dropdown_view
-
 /**
  * this view lets the user (de-)select aspect memberships in the context
  * of another users profile or the contact page.
@@ -7,13 +5,15 @@
  * updates to the list of aspects are immediately propagated to the server, and
  * the results are dislpayed as flash messages.
  */
-app.views.AspectMembership = app.views.AspectsDropdown.extend({
-
-  events: {
-    "click ul.aspect_membership.dropdown-menu > li.aspect_selector": "_clickHandler"
-  },
+app.views.AspectMembershipBlueprint = Backbone.View.extend({
 
   initialize: function() {
+    // attach event handler, removing any previous instances
+    var selector = '.dropdown.aspect_membership .dropdown_list > li';
+    $('body')
+      .off('click', selector)
+      .on('click', selector, _.bind(this._clickHandler, this));
+
     this.list_item = null;
     this.dropdown  = null;
   },
@@ -22,7 +22,7 @@ app.views.AspectMembership = app.views.AspectsDropdown.extend({
   //   -> addMembership
   //   -> removeMembership
   _clickHandler: function(evt) {
-    this.list_item = $(evt.target).closest('li.aspect_selector');
+    this.list_item = $(evt.target);
     this.dropdown  = this.list_item.parent();
 
     this.list_item.addClass('loading');
@@ -72,16 +72,17 @@ app.views.AspectMembership = app.views.AspectsDropdown.extend({
     }
 
     li.attr('data-membership_id', membership_id) // just to be sure...
-      .data('membership_id', membership_id);
+      .data('membership_id', membership_id)
+      .addClass('selected');
 
-    this.updateSummary(li);
+    this.updateSummary();
     this._done();
   },
 
   // show an error flash msg
   _displayError: function(msg_id) {
     this._done();
-    this.dropdown.closest('.aspect_membership_dropdown').removeClass('open'); // close the dropdown
+    this.dropdown.removeClass('active'); // close the dropdown
 
     var msg = Diaspora.I18n.t(msg_id, { 'name': this._name() });
     Diaspora.page.flashMessages.render({ 'success':false, 'notice':msg });
@@ -106,8 +107,8 @@ app.views.AspectMembership = app.views.AspectsDropdown.extend({
     var li = this.dropdown.find('li[data-membership_id="'+membership_id+'"]');
 
     li.removeAttr('data-membership_id')
-      .removeData('membership_id');
-    this.updateSummary(li);
+      .removeData('membership_id')
+      .removeClass('selected');
 
     // we just removed the last aspect, inform the user with a flash message
     // that he is no longer sharing with that person
@@ -116,6 +117,7 @@ app.views.AspectMembership = app.views.AspectsDropdown.extend({
       Diaspora.page.flashMessages.render({ 'success':true, 'notice':msg });
     }
 
+    this.updateSummary();
     this._done();
   },
 
@@ -127,8 +129,33 @@ app.views.AspectMembership = app.views.AspectsDropdown.extend({
   },
 
   // refresh the button text to reflect the current aspect selection status
-  updateSummary: function(target) {
-    this._toggleCheckbox(target);
-    this._updateButton('green');
+  updateSummary: function() {
+    var btn = this.dropdown.parents('div.aspect_membership').find('.button.toggle');
+    var aspects_cnt = this.dropdown.find('li.selected').length;
+    var txt;
+
+    if( aspects_cnt == 0 ) {
+      btn.removeClass('in_aspects');
+      txt = Diaspora.I18n.t('aspect_dropdown.toggle.zero');
+    } else {
+      btn.addClass('in_aspects');
+      txt = this._pluralSummaryTxt(aspects_cnt);
+    }
+
+    btn.text(txt + ' ▼');
   },
+
+  _pluralSummaryTxt: function(cnt) {
+    var all_aspects_cnt = this.dropdown.find('li').length;
+
+    if( cnt == 1 ) {
+      return this.dropdown.find('li.selected').first().text();
+    }
+
+    if( cnt == all_aspects_cnt ) {
+      return Diaspora.I18n.t('aspect_dropdown.all_aspects');
+    }
+
+    return Diaspora.I18n.t('aspect_dropdown.toggle', { 'count':cnt.toString() });
+  }
 });