+++ app/helpers/aspect_global_helper.rb
@@ -3,7 +3,7 @@
 #   the COPYRIGHT file.
 
 module AspectGlobalHelper
-  def aspect_membership_dropdown(contact, person, hang, aspect=nil)
+  def aspect_membership_dropdown(contact, person, hang, aspect=nil, force_bootstrap=false)
     aspect_membership_ids = {}
 
     selected_aspects = all_aspects.select{|aspect| contact.in_aspect?(aspect)}
@@ -12,12 +12,21 @@ module AspectGlobalHelper
       aspect_membership_ids[a.id] = record.id
     end
 
-    render "shared/aspect_dropdown",
-      :selected_aspects => selected_aspects,
-      :aspect_membership_ids => aspect_membership_ids,
-      :person => person,
-      :hang => hang,
-      :dropdown_class => "aspect_membership"
+    if bootstrap? || force_bootstrap
+      render "aspect_memberships/aspect_membership_dropdown",
+        :selected_aspects => selected_aspects,
+        :aspect_membership_ids => aspect_membership_ids,
+        :person => person,
+        :hang => hang,
+        :dropdown_class => "aspect_membership"
+    else
+      render "aspect_memberships/aspect_membership_dropdown_blueprint",
+        :selected_aspects => selected_aspects,
+        :aspect_membership_ids => aspect_membership_ids,
+        :person => person,
+        :hang => hang,
+        :dropdown_class => "aspect_membership"
+    end
   end
 
   def aspect_dropdown_list_item(aspect, am_id=nil)